Fix building on musl
* Add missing linux/stddef.h includes
* Explicitly cast PAGE_SIZE to size_t. PAGE_SIZE is defined without UL
  suffix in musl
* Musl doesn't define PTHREAD_RECURSIVE_MUTEX_INITIALIZER_NP, so
  initialize the mutexes with pthread_once.

+++ include/linux/slab.h
@@ -22,7 +22,7 @@ static inline void *kmalloc(size_t size, gfp_t flags)
 
 	p = size
 	    ? aligned_alloc(min(rounddown_pow_of_two(size),
-				PAGE_SIZE), size)
+				(size_t)PAGE_SIZE), size)
 	    : malloc(0);
 	if (p && (flags & __GFP_ZERO))
 		memset(p, 0, size);
